remove flicker from plot
