+++ src/__test__/components/plots/TrajectoryAnalysisPlot.test.jsx
@@ -9,18 +9,17 @@ import TrajectoryAnalysisPlot from 'components/plots/TrajectoryAnalysisPlot';
 
 import { initialPlotConfigStates } from 'redux/reducers/componentConfig/initialState';
 
-import mockAPI, { generateDefaultMockAPIResponses, statusResponse, delayedResponse } from '__test__/test-utils/mockAPI';
+import mockAPI, { generateDefaultMockAPIResponses, promiseResponse } from '__test__/test-utils/mockAPI';
 import createTestComponentFactory from '__test__/test-utils/testComponentFactory';
 import fake from '__test__/test-utils/constants';
 import mockEmbedding from '__test__/data/embedding.json';
 import { seekFromS3 } from 'utils/work/seekWorkResponse';
 import { loadBackendStatus } from 'redux/actions/backendStatus';
-import WorkResponseError from 'utils/http/errors/WorkResponseError';
 
 import mockStartingNodes from '__test__/data/starting_nodes.json';
+import mockProcessingConfig from '__test__/data/processing_config.json';
 
 import { plotTypes } from 'utils/constants';
-import userEvent from '@testing-library/user-event';
 
 enableFetchMocks();
 
@@ -118,86 +117,23 @@ describe('Trajectory analysis plot', () => {
     });
   });
 
-  it('Shows a loader if there is no config', async () => {
-    await renderTrajectoryAnalysisPlot(storeState, { config: null });
+  it('Shows warning if embedding method is tsne', async () => {
+    const tsneSettings = { ...mockProcessingConfig };
+    tsneSettings.processingConfig.configureEmbedding.embeddingSettings.method = 'tsne';
 
-    expect(screen.getByText(/We're getting your data/i)).toBeInTheDocument();
-    await waitFor(async () => {
-      await expect(screen.queryByRole('graphics-document', { name: defaultShownPlotDescription })).toBeNull();
-    });
-  });
-
-  it('Shows a loader if cell sets is loading', async () => {
-    const cellSetErrorResponse = {
+    fetchMock.resetMocks();
+    fetchMock.mockIf(/.*/, mockAPI({
       ...defaultAPIResponse,
-      [`experiments/${experimentId}/cellSets`]: () => delayedResponse({ body: 'Not found', status: 404 }, 4000),
-    };
-
-    fetchMock.mockIf(/.*/, mockAPI(cellSetErrorResponse));
-
-    await renderTrajectoryAnalysisPlot(storeState);
-
-    expect(screen.getByText(/We're getting your data/i)).toBeInTheDocument();
-    await waitFor(async () => {
-      await expect(screen.queryByRole('graphics-document', { name: defaultShownPlotDescription })).toBeNull();
-    });
-  });
-
-  it('Shows an error if fetching cell sets throw an error', async () => {
-    const cellSetErrorResponse = {
-      ...defaultAPIResponse,
-      [`experiments/${experimentId}/cellSets`]: () => statusResponse(500, 'some random error'),
-    };
-
-    fetchMock.mockIf(/.*/, mockAPI(cellSetErrorResponse));
+      [`experiments/${experimentId}/processingConfig`]: () => promiseResponse(
+        JSON.stringify(tsneSettings),
+      ),
+    }));
 
     await renderTrajectoryAnalysisPlot(storeState);
 
-    expect(screen.getByText(/We're sorry, we couldn't load this/i)).toBeInTheDocument();
     await waitFor(async () => {
-      await expect(screen.queryByRole('graphics-document', { name: defaultShownPlotDescription })).toBeNull();
+      expect(screen.getByText(/The embedding and trajectory below are generated from a UMAP embedding of your data/)).toBeInTheDocument();
+      expect(screen.getByRole('graphics-document', { name: defaultShownPlotDescription })).toBeInTheDocument();
     });
-  });
-
-  it('Shows a loader if embedding data is loading', async () => {
-    seekFromS3
-      .mockReset()
-      .mockImplementationOnce(() => null)
-      .mockImplementationOnce(() => delayedResponse({ body: 'Not found', status: 404 }, 4000));
-
-    await renderTrajectoryAnalysisPlot(storeState);
-
-    expect(screen.getByText(/We're getting your data/i)).toBeInTheDocument();
-    await waitFor(async () => {
-      await expect(screen.queryByRole('graphics-document', { name: defaultShownPlotDescription })).toBeNull();
-    });
-  });
-
-  it('Shows an error if there is an error fetching embedding', async () => {
-    seekFromS3
-      .mockReset()
-      .mockImplementationOnce(() => null)
-      .mockImplementationOnce(() => Promise.reject(new WorkResponseError('some random error')));
-
-    await renderTrajectoryAnalysisPlot(storeState);
-
-    expect(screen.getByText(/We had an error on our side while we were completing your request/i)).toBeInTheDocument();
-    await waitFor(async () => {
-      await expect(screen.queryByRole('graphics-document', { name: defaultShownPlotDescription })).toBeNull();
-    });
-  });
-
-  it('Shows an error if there is an error fetching plot data', async () => {
-    await renderTrajectoryAnalysisPlot(storeState, { plotDataError: true });
-
-    expect(screen.getByText(/We're sorry, we couldn't load this./i)).toBeInTheDocument();
-    await waitFor(async () => {
-      await expect(screen.queryByRole('graphics-document', { name: defaultShownPlotDescription })).toBeNull();
-    });
-
-    // Clicking retry will fire onPlotDataErroRetry
-    userEvent.click(screen.getByText(/Try again/));
-
-    expect(mockOnPlotDataErrorRetry).toHaveBeenCalledTimes(1);
   });
 });
